DR2-1148 feat: Lambda for parsing documents from TRE.
The lambda does the following.

* Reads a TRE input message. It's expecting the raw json, not an SQS or
  SNS message
* Downloads the package from TRE.
* Unzips it
* Untars it
* Uploads all files from the package to S3 with a UUID. This includes
  files we don't care about but it's easier than trying to parse json on
  the fly
* Parses the metadata json from the package.
* Gets a series and department code from a static map based on the cite.
* Generates the bagit files and uploads them to S3 in memory.
* Copies the docx and metadata files into the data/ directory
* Starts a step function execution with the judgment details.

The department and series lookup is very judgment specific but this can
be changed if we start taking in other transfers.

The tests are split into two types:

LambdaTest - These are "integration" tests. They use wiremock to mock
the responses from AWS but otherwise use the real clients.
FileProcessorTest and SeriesMapperTest. These are more "unit" tests in
that the clients are mocked.

This PR depends on https://github.com/nationalarchives/da-aws-clients/pull/25
The tests won't pass until that is deployed as we need the client
methods.