DR2-1197 Remove Press Summary of from the title.
There's a few corrections that needed to be made here.
The asset and folder title were taken from the docx filename without the
docx. This has been changed now to:
* For a folder use the name from the TRE json with "Press Summary of "
  removed. If this name is missing, use the docx filename.
* For an asset, do the same but don't strip out "Press summary of "

I've updated the tests to do a loop around the possible titles.

It looks like a lot of changes but it's mostly whitespace.