DR2-1416 Improved logging.
This adds some extra logging to the lambda to make it easier to see what
is being processed.

The map that is passed as the first argument to logger.info is sent to
Cloudwatch as a json object that can be parsed in queries.
The first two log lines have only the batch ID as we don't yet have the
citeable reference from TRE.

The remainder of the log lines contain the batch ID and the citeable
reference. I've called it fileReference here as that's what TRE have
called it so it keeps it all consistent.

This logging is *not* meant to contain a comprehensive account of all
IDs that are passing through the system, for example, it says "Copied
bagit files to $outputBucket" The information as to exactly what they
are is in S3 and doesn't belong here. The focus is on making the logs
short and readable as guide to where a particular ingest is at the time.