DR2-1352 Use batch id for execution name.
We were originally going to let Preservica handle duplicate assets but
it looks like this doesn't work.

As a temporary solution, I've changed the step function execution name
to be just the batch id. This means that if a duplicate batch id comes
in, it will fail because the name already exists.